Updates concerning PhotData
Correcting julia version in Project.toml to ≥0.7.
In MCMphotolysis.jl, exporting PhotData, moving __init__ function to the
beginning before include files, updating docstrings and function
j_parameters to get fitting parameters for the updated l parameter.
In function getTUVdata in rdfiles.jl, removing the calculation of the
order of magnitude as it is outsourced to module filehandling.
In function fit_jold in fitTUV.jl, updated initial guesses of l to
include exp(-0.3) in the guess. Updated treatment of order of magnitude,
needs to be removed temporarily as original values are now save in
PhotData. Updates in getMCMparams to new data structure PhotData.