Text file output for j_parameters
In output.jl, new function wrt_newparams to write parameters and errors
to a formatted output file and unformatted to a csv file. Referencing
LsgFit in the header of the formatted files in wrt_params and
wrt_newparams.
Using abs to calculate sigmal in function fitl in fitTUV.jl to avoid a
negative standard deviation when dividing through a negative l param.
New function call to wrt_newparams in j_parameters in MCMphotolysis.jl.
Adding csv files to gitignore.