Treat TUVdata.order as Int
Update packages in Manifest.toml. This includes updates in filehandling
to change TUVdata.order to a vector of integers.
Remove conversions of order to Int for plotting in output.jl.
Ensure base 10 is an integers in conversions, when excluding or
including order of magnitude in fitTUV.jl and output.jl.