Add MCM and TUV reaction numbers to output, fix
Update dependencies in Manifest.toml including update of filehandling to
version 0.3.1, which return additional fields mcm and tuv in TUVdata
with the respective reaction numbers.
In output.jl, rename function beautify_chem to set_titles. Now all
TUVdata is handed and MCM reaction numbers are added to the titles.
Adjusted calls of plotting functions in output.jl. Write MCM reaction
numbers to formatted output files and MCM and TUV numbers to the csv
file produced by j_parameters.
In j_oldpars, call wrt_params before plot_jold as plot_jold is the time
limiting step, so parameters.dat can already be inspected, when plots
are produced.
FIX: When calculating ss_tot for the calculation of R2, reduce it by the
     order of magnitude as it was done for the other fit results.