Fix #4, use Inf for errors after fit fail
In function convergel in fitTUV.jl, remove index i from lpar, when
trying to remove lowest or highest ozone columns as lpar is passed as
vector and not as matrix. Fixes #4.
In function fitl in fitTUV.jl, assign array of Inf to sigmal and skip
calling standard_error. This avoids the abortion of the script, if l
fails to converge.
Export StatData in MCMphotolysis.jl.
Update dependencies in Manifest.toml. 
Set version to v0.2.1-DEV in Project.toml.